+++ src/data/configs.js
@@ -72,16 +72,16 @@ var configs = {
             {
                 'buttons': [
                     { 'id': 'openInFgTab' },
-                    { 'id': 'copyLinkText' },
-                    { 'id': 'openInBgTab' },
                     { 'id': 'copyUrl' },
+                    { 'id': 'openInBgTab' },
+                    { 'id': 'copyLinkText' },
                 ]
             }
         ]
     },
 
     'selectionMenu': {
-        'color': "#000000",
+        'color': "#B44141",
         'rockerAction': 'copyText',
         'mouseWheelUpAction': 'noAction',
         'mouseWheelDownAction': 'noAction',
@@ -98,7 +98,7 @@ var configs = {
 
     'textFieldMenu': {
         'width': 100,
-        'color': "#000000",
+        'color': "#499447",
         'rockerAction': 'pasteText',
         'mouseWheelUpAction': 'noAction',
         'mouseWheelDownAction': 'noAction',
@@ -106,17 +106,18 @@ var configs = {
             {
                 'buttons': [
                     { 'id': 'copyText' },
-                    { 'id': 'selectAllText' },
-                    { 'id': 'cutText' },
                     { 'id': 'pasteText' },
+                    { 'id': 'cutText' },
+                    { 'id': 'selectAllText' },
+                    { 'id': 'clearInputField' },
                 ]
             }
         ]
     },
 
     'imageMenu': {
-        // 'color': "#c69a15",
-        'color': "#d80000",
+        // 'color': "#d80000",
+        'color': "#F48024",
         'rockerAction': 'openInFgTab',
         'levels': [
             {